Refactor MeshBenchmarksRunner to reduce the number of explicitly instantiated templates
The instances for different cell topology and mesh config template, but
different Real, GlobalIndex, and LocalIndex types, still generated a lot
of common codes for the underlying data structures and algorithms
(especially sparse matrices and vectors). Hence, the splitting was not
efficient. We were also hitting strange segfaults in the nvcc's host
code implementation of extended lambda functions, probably because some
duplicate code was not linked together correctly.